fix bug in get enrollment report

+++ src/main/java/br/com/alura/school/support/dto/EnrollmentReportDto.java
@@ -1,20 +1,20 @@
 package br.com.alura.school.support.dto;
 
 public class EnrollmentReportDto {
-    private Integer qtdEnrollments;
+    private Long qtdEnrollments;
 
     private String email;
 
-    public EnrollmentReportDto(Integer qtdEnrollments, String email) {
+    public EnrollmentReportDto(Long qtdEnrollments, String email) {
         this.qtdEnrollments = qtdEnrollments;
         this.email = email;
     }
 
-    public Integer getQtdEnrollments() {
+    public Long getQtdEnrollments() {
         return qtdEnrollments;
     }
 
-    public void setQtdEnrollments(Integer qtdEnrollments) {
+    public void setQtdEnrollments(Long qtdEnrollments) {
         this.qtdEnrollments = qtdEnrollments;
     }
